+++ changelog.md
@@ -78,4 +78,5 @@
 
 ## v4.1.1
 - added modal to pit results view, for better viewing on mobile devices
+- fixed image preview breaking when re-uploading (used 'parentNode.remove(element)` instead of `element.remove`)
 - moved `match.module.css`